requirements.txt
langgraph==0.2.16
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
openai==1.58.1
python-dotenv==1.0.1
pillow==10.4.0
pydantic==2.8.2
typing-extensions==4.12.2 